requirements.txt
Django
mysqlclient
django-crispy-forms
crispy-bootstrap5